Fixes around region painting
